chore: create a separate stream package (#8602)
