+++ core/internal/stream/handler.go
@@ -1,4 +1,4 @@
-package server
+package stream
 
 import (
 	"errors"
