+++ core/internal/stream/handler_test.go
@@ -1,4 +1,4 @@
-package server_test
+package stream_test
 
 import (
 	"fmt"
@@ -8,8 +8,8 @@ import (
 	"github.com/stretchr/testify/assert"
 	"github.com/wandb/wandb/core/internal/observability"
 	"github.com/wandb/wandb/core/internal/runwork"
+	"github.com/wandb/wandb/core/internal/stream"
 	"github.com/wandb/wandb/core/internal/version"
-	"github.com/wandb/wandb/core/pkg/server"
 	spb "github.com/wandb/wandb/core/pkg/service_go_proto"
 )
 
@@ -17,10 +17,10 @@ func makeHandler(
 	inChan, fwdChan chan runwork.Work,
 	outChan chan *spb.Result,
 	commit string,
-) *server.Handler {
-	h := server.NewHandler(
+) *stream.Handler {
+	h := stream.NewHandler(
 		commit,
-		server.HandlerParams{
+		stream.HandlerParams{
 			Logger:          observability.NewNoOpLogger(),
 			Settings:        &spb.Settings{},
 			FwdChan:         fwdChan,
@@ -676,9 +676,9 @@ func TestHandlePartialHistory(t *testing.T) {
 
 	for _, tc := range testCases {
 		t.Run(tc.name, func(t *testing.T) {
-			inChan := make(chan runwork.Work, server.BufferSize)
-			fwdChan := make(chan runwork.Work, server.BufferSize)
-			outChan := make(chan *spb.Result, server.BufferSize)
+			inChan := make(chan runwork.Work, stream.BufferSize)
+			fwdChan := make(chan runwork.Work, stream.BufferSize)
+			outChan := make(chan *spb.Result, stream.BufferSize)
 
 			makeHandler(inChan, fwdChan, outChan, "" /*commit*/)
 
@@ -776,9 +776,9 @@ func TestHandleHistory(t *testing.T) {
 	}
 	for _, tc := range testCases {
 		t.Run(tc.name, func(t *testing.T) {
-			inChan := make(chan runwork.Work, server.BufferSize)
-			fwdChan := make(chan runwork.Work, server.BufferSize)
-			outChan := make(chan *spb.Result, server.BufferSize)
+			inChan := make(chan runwork.Work, stream.BufferSize)
+			fwdChan := make(chan runwork.Work, stream.BufferSize)
+			outChan := make(chan *spb.Result, stream.BufferSize)
 
 			makeHandler(inChan, fwdChan, outChan, "" /*commit*/)
