+++ core/internal/stream/headfoot.go
@@ -1,5 +1,5 @@
 // This file should be depreacted and removed in the future, in favor of printing the footer in the client
-package server
+package stream
 
 import (
 	"fmt"
