+++ core/internal/stream/sender_test.go
@@ -1,6 +1,7 @@
-package server_test
+package stream_test
 
 import (
+	"context"
 	"testing"
 
 	"github.com/Khan/genqlient/graphql"
@@ -12,8 +13,9 @@ import (
 	"github.com/wandb/wandb/core/internal/observability"
 	"github.com/wandb/wandb/core/internal/runworktest"
 	wbsettings "github.com/wandb/wandb/core/internal/settings"
+	"github.com/wandb/wandb/core/internal/stream"
 	"github.com/wandb/wandb/core/internal/watchertest"
-	"github.com/wandb/wandb/core/pkg/server"
+	"github.com/wandb/wandb/core/pkg/artifacts"
 	spb "github.com/wandb/wandb/core/pkg/service_go_proto"
 	"google.golang.org/protobuf/types/known/wrapperspb"
 )
@@ -36,7 +38,7 @@ const validLinkArtifactResponse = `{
 	"linkArtifact": { "versionIndex": 0 }
 }`
 
-func makeSender(client graphql.Client, resultChan chan *spb.Result) *server.Sender {
+func makeSender(client graphql.Client, resultChan chan *spb.Result) *stream.Sender {
 	runWork := runworktest.New()
 	logger := observability.NewNoOpLogger()
 	settings := wbsettings.From(&spb.Settings{
@@ -44,8 +46,8 @@ func makeSender(client graphql.Client, resultChan chan *spb.Result) *server.Send
 		Console: &wrapperspb.StringValue{Value: "off"},
 		ApiKey:  &wrapperspb.StringValue{Value: "test-api-key"},
 	})
-	backend := server.NewBackend(logger, settings)
-	fileStream := server.NewFileStream(
+	backend := stream.NewBackend(logger, settings)
+	fileStream := stream.NewFileStream(
 		backend,
 		logger,
 		nil, // operations
@@ -53,12 +55,12 @@ func makeSender(client graphql.Client, resultChan chan *spb.Result) *server.Send
 		settings,
 		nil, // peeker
 	)
-	fileTransferManager := server.NewFileTransferManager(
+	fileTransferManager := stream.NewFileTransferManager(
 		filetransfer.NewFileTransferStats(),
 		logger,
 		settings,
 	)
-	runfilesUploader := server.NewRunfilesUploader(
+	runfilesUploader := stream.NewRunfilesUploader(
 		runWork,
 		logger,
 		nil, // operations
@@ -68,9 +70,9 @@ func makeSender(client graphql.Client, resultChan chan *spb.Result) *server.Send
 		watchertest.NewFakeWatcher(),
 		client,
 	)
-	sender := server.NewSender(
+	sender := stream.NewSender(
 		runWork,
-		server.SenderParams{
+		stream.SenderParams{
 			Logger:              logger,
 			Settings:            settings,
 			Backend:             backend,
@@ -287,4 +289,135 @@ func TestSendUseArtifact(t *testing.T) {
 		},
 	}
 	sender.SendRecord(useArtifact)
+}
+
+var validFetchOrgEntityFromEntityResponse = `{
+	"entity": {
+		"organization": {
+			"name": "orgName",
+			"orgEntity": {
+				"name": "orgEntityName_123"
+			}
+		}
+	}
+}`
+
+func TestLinkRegistryArtifact(t *testing.T) {
+	registryProject := artifacts.RegistryProjectPrefix + "projectName"
+	expectLinkArtifactFailure := "expect link artifact to fail, wrong org entity"
+
+	testCases := []struct {
+		name              string
+		inputOrganization string
+		isOldServer       bool
+		errorMessage      string
+	}{
+		{"Link registry artifact with orgName updated server", "orgName", false, ""},
+		{"Link registry artifact with orgName old server", "orgName", true, expectLinkArtifactFailure},
+		{"Link registry artifact with orgEntity name updated server", "orgEntityName_123", false, ""},
+		{"Link registry artifact with orgEntity name old server", "orgEntityName_123", true, ""},
+		{"Link registry artifact with short hand path updated server", "", false, ""},
+		{"Link registry artifact with short hand path old server", "", true, "unsupported"},
+		{"Link with wrong org/orgEntity name with updated server", "potato", false, "update the target path"},
+		{"Link with wrong org/orgEntity name with updated server", "potato", true, expectLinkArtifactFailure},
+	}
+	for _, tc := range testCases {
+		mockGQL := gqlmock.NewMockClient()
+
+		newLinker := func(req *spb.LinkArtifactRequest) *artifacts.ArtifactLinker {
+			return &artifacts.ArtifactLinker{
+				Ctx:           context.Background(),
+				LinkArtifact:  req,
+				GraphqlClient: mockGQL,
+			}
+		}
+
+		// If user is on old server, we can't fetch the org entity name so just directly call link artifact
+		numExpectedRequests := 3
+		if tc.isOldServer {
+			numExpectedRequests = 2
+		}
+
+		t.Run("Link registry artifact with orgName updated server", func(t *testing.T) {
+			req := &spb.LinkArtifactRequest{
+				ClientId:              "clientId123",
+				PortfolioName:         "portfolioName",
+				PortfolioEntity:       "entityName",
+				PortfolioProject:      registryProject,
+				PortfolioAliases:      nil,
+				PortfolioOrganization: tc.inputOrganization,
+			}
+
+			var validTypeFieldsResponse string
+			if tc.isOldServer {
+				validTypeFieldsResponse = `{"TypeInfo": {"fields": []}}`
+			} else {
+				validTypeFieldsResponse = `{
+		"TypeInfo": {
+			"fields": [{"name": "orgEntity"}]
+		}
+	}`
+			}
+			mockGQL.StubMatchOnce(
+				gqlmock.WithOpName("TypeFields"),
+				validTypeFieldsResponse,
+			)
+
+			mockGQL.StubMatchOnce(
+				gqlmock.WithOpName("LinkArtifact"),
+				validLinkArtifactResponse,
+			)
+
+			mockGQL.StubMatchOnce(
+				gqlmock.WithOpName("FetchOrgEntityFromEntity"),
+				validFetchOrgEntityFromEntityResponse,
+			)
+
+			linker := newLinker(req)
+			err := linker.Link()
+			if err != nil {
+				assert.NotEmpty(t, tc.errorMessage)
+				assert.ErrorContainsf(t, err, tc.errorMessage,
+					"Expected error containing: %s", tc.errorMessage)
+				return
+			}
+
+			// This error is not triggered by Link() because its linkArtifact that fails
+			// and we aren't actually calling it.
+			// Here we are checking that the org entity being passed into linkArtifact
+			// is wrong so we know the query will fail.
+			if tc.errorMessage == expectLinkArtifactFailure {
+				requests := mockGQL.AllRequests()
+				assert.Len(t, requests, numExpectedRequests)
+
+				// Confirms that the request is incorrectly put into link artifact graphql request
+				gqlmock.AssertRequest(t,
+					gqlmock.WithVariables(
+						gqlmock.GQLVar("projectName", gomock.Eq(registryProject)),
+						// Here the entity name is not orgEntityName_123 and this will fail if actually called
+						gqlmock.GQLVar("entityName", gomock.Not(gomock.Eq("orgEntityName_123"))),
+						gqlmock.GQLVar("artifactPortfolioName", gomock.Eq("portfolioName")),
+						gqlmock.GQLVar("clientId", gomock.Eq("clientId123")),
+						gqlmock.GQLVar("artifactId", gomock.Nil()),
+					),
+					requests[numExpectedRequests-1])
+			} else {
+				// If no error, check that we are passing in the correct org entity name into linkArtifact
+				assert.Empty(t, tc.errorMessage)
+				assert.NoError(t, err)
+				requests := mockGQL.AllRequests()
+				assert.Len(t, requests, numExpectedRequests)
+
+				gqlmock.AssertRequest(t,
+					gqlmock.WithVariables(
+						gqlmock.GQLVar("projectName", gomock.Eq(registryProject)),
+						gqlmock.GQLVar("entityName", gomock.Eq("orgEntityName_123")),
+						gqlmock.GQLVar("artifactPortfolioName", gomock.Eq("portfolioName")),
+						gqlmock.GQLVar("clientId", gomock.Eq("clientId123")),
+						gqlmock.GQLVar("artifactId", gomock.Nil()),
+					),
+					requests[numExpectedRequests-1])
+			}
+		})
+	}
 }
