+++ core/internal/stream/store.go
@@ -1,4 +1,4 @@
-package server
+package stream
 
 import (
 	"bytes"
