+++ core/pkg/server/connection.go
@@ -15,6 +15,7 @@ import (
 	"github.com/wandb/wandb/core/internal/observability"
 	"github.com/wandb/wandb/core/internal/sentry_ext"
 	"github.com/wandb/wandb/core/internal/settings"
+	"github.com/wandb/wandb/core/internal/stream"
 
 	spb "github.com/wandb/wandb/core/pkg/service_go_proto"
 	"google.golang.org/protobuf/proto"
@@ -27,7 +28,7 @@ const (
 )
 
 type ConnectionOptions struct {
-	StreamMux    *StreamMux
+	StreamMux    *stream.StreamMux
 	Conn         net.Conn
 	SentryClient *sentry_ext.Client
 	Commit       string
@@ -53,7 +54,7 @@ type Connection struct {
 
 	// A map that associates stream IDs with active streams (or runs). This helps
 	// track the streams associated with this connection.
-	streamMux *StreamMux
+	streamMux *stream.StreamMux
 
 	// id is the unique id for the connection
 	id string
@@ -71,7 +72,7 @@ type Connection struct {
 	// The stream associated with this connection. While each connection has one
 	// stream, a stream can have multiple active connections, typically for a
 	// multi-client session.
-	stream *Stream
+	stream *stream.Stream
 
 	// The current W&B Git commit hash, identifying the specific version of the binary.
 	commit string
@@ -318,14 +319,14 @@ func (nc *Connection) handleInformInit(msg *spb.ServerInformInitRequest) {
 		sentryClient = nc.sentryClient
 	}
 
-	nc.stream = NewStream(
-		StreamOptions{
+	nc.stream = stream.NewStream(
+		stream.StreamOptions{
 			Commit:     nc.commit,
 			Settings:   settings,
 			Sentry:     sentryClient,
 			LoggerPath: nc.loggerPath,
 		})
-	nc.stream.AddResponders(ResponderEntry{nc, nc.id})
+	nc.stream.AddResponders(stream.ResponderEntry{Responder: nc, ID: nc.id})
 	nc.stream.Start()
 	slog.Info("handleInformInit: stream started", "streamId", streamId, "id", nc.id)
 
@@ -345,15 +346,15 @@ func (nc *Connection) handleInformInit(msg *spb.ServerInformInitRequest) {
 func (nc *Connection) handleInformStart(msg *spb.ServerInformStartRequest) {
 	// todo: if we keep this and end up updating the settings here
 	//       we should update the stream logger to use the new settings as well
-	nc.stream.settings = settings.From(msg.GetSettings())
+	nc.stream.Settings = settings.From(msg.GetSettings())
 
 	// update sentry tags
 	// add attrs from settings:
-	nc.stream.logger.SetGlobalTags(observability.Tags{
-		"run_url": nc.stream.settings.GetRunURL(),
+	nc.stream.Logger.SetGlobalTags(observability.Tags{
+		"run_url": nc.stream.Settings.GetRunURL(),
 	})
 	// TODO: remove this once we have a better observability setup
-	nc.stream.logger.CaptureInfo("wandb-core", nil)
+	nc.stream.Logger.CaptureInfo("wandb-core", nil)
 }
 
 // handleInformAttach handles the new connection attaching to an existing stream.
@@ -370,14 +371,14 @@ func (nc *Connection) handleInformAttach(msg *spb.ServerInformAttachRequest) {
 	if err != nil {
 		slog.Error("handleInformAttach: stream not found", "streamId", streamId, "id", nc.id)
 	} else {
-		nc.stream.AddResponders(ResponderEntry{nc, nc.id})
+		nc.stream.AddResponders(stream.ResponderEntry{Responder: nc, ID: nc.id})
 		// TODO: we should redo this attach logic, so that the stream handles
 		//       the attach logic
 		resp := &spb.ServerResponse{
 			ServerResponseType: &spb.ServerResponse_InformAttachResponse{
 				InformAttachResponse: &spb.ServerInformAttachResponse{
 					XInfo:    msg.XInfo,
-					Settings: nc.stream.settings.Proto,
+					Settings: nc.stream.Settings.Proto,
 				},
 			},
 		}
@@ -406,8 +407,8 @@ func (nc *Connection) handleAuthenticate(msg *spb.ServerAuthenticateRequest) {
 			BaseUrl: &wrapperspb.StringValue{Value: msg.BaseUrl},
 		},
 	}
-	backend := NewBackend(observability.NewNoOpLogger(), s) // TODO: use a real logger
-	graphqlClient := NewGraphQLClient(backend, s, &observability.Peeker{})
+	backend := stream.NewBackend(observability.NewNoOpLogger(), s) // TODO: use a real logger
+	graphqlClient := stream.NewGraphQLClient(backend, s, &observability.Peeker{})
 
 	data, err := gql.Viewer(context.Background(), graphqlClient)
 	if err != nil || data == nil || data.GetViewer() == nil || data.GetViewer().GetEntity() == nil {
